+++ packages/cascader/src/popup/search-popup.tsx
@@ -2,8 +2,7 @@ import React, { useEffect, useState, CSSProperties, useRef } from "react"
 import isEqual from "react-fast-compare"
 import { Empty } from "@illa-design/empty"
 import { Checkbox } from "@illa-design/checkbox"
-import Store from "../store"
-import { Node } from "../node"
+import { Node, Store } from "../node"
 import { OptionProps } from "../interface"
 import { applyOptionStyle, optionListStyle, searchListWrapper } from "./style"
